Fix cmap/kern parser validation gaps and integer overflows (PR review round 2)
- CmapTable format-4: reject subtables whose segments are not strictly
  increasing/non-overlapping by endCode, whose startCode exceeds endCode, or
  whose final segment is not the mandatory 0xFFFF terminator.
- CmapTable format-4: widen the indirect idRangeOffset glyph-index address
  computation to long arithmetic and reject addresses that point backward
  (before glyphIdArray) or past the subtable end, instead of only bounds-
  checking the upper side with unwidened int arithmetic.
- CmapTable format-12: reject subtables whose groups are not strictly
  increasing/non-overlapping by (startCharCode, endCharCode).
- KernTable format-0: widen the body-offset bounds check to long arithmetic
  to match the other overflow-safe checks already in this file.
- Add regression tests for all four fixes (CmapTableTests, KernTableTests),
  each verified to fail against the pre-fix code and pass post-fix.

+++ src/DemaConsulting.CanvasNet/Fonts/KernTable.cs
@@ -154,7 +154,7 @@ internal sealed class KernTable
     /// </summary>
     private static (ushort, ushort, short)[]? TryParseFormat0(byte[] data, int bodyOffset, int subtableEnd)
     {
-        if (bodyOffset + 8 > subtableEnd)
+        if ((long)bodyOffset + 8 > subtableEnd)
         {
             return null;
         }
